// packages/api/src/codex-protocol/v2/RateLimitSnapshot.ts
// GENERATED CODE! DO NOT MODIFY BY HAND!

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import { type PlanType } from "../PlanType";
import { type CreditsSnapshot } from "./CreditsSnapshot";
import { type RateLimitReachedType } from "./RateLimitReachedType";
import { type RateLimitWindow } from "./RateLimitWindow";

export type RateLimitSnapshot = {
  limitId: string | null;
  limitName: string | null;
  primary: RateLimitWindow | null;
  secondary: RateLimitWindow | null;
  credits: CreditsSnapshot | null;
  planType: PlanType | null;
  rateLimitReachedType: RateLimitReachedType | null;
};